MED-442: - move the ADs Provides to files in a special folders
